Extract Multipart store methods to MultipartStore
