+++ server/src/main/java/com/adobe/testing/s3mock/S3MockConfiguration.java
@@ -191,7 +191,8 @@ class S3MockConfiguration implements WebMvcConfigurer {
      */
     @ExceptionHandler(S3Exception.class)
     public ResponseEntity<ErrorResponse> handleS3Exception(final S3Exception s3Exception) {
-      LOG.info("Responding with status {}: {}", s3Exception.getStatus(), s3Exception.getMessage());
+      LOG.info("Responding with status {}: {}", s3Exception.getStatus(), s3Exception.getMessage(),
+          s3Exception);
 
       final ErrorResponse errorResponse = new ErrorResponse();
       errorResponse.setCode(s3Exception.getCode());
@@ -224,7 +225,8 @@ class S3MockConfiguration implements WebMvcConfigurer {
      */
     @ExceptionHandler(IllegalStateException.class)
     public ResponseEntity<ErrorResponse> handleS3Exception(IllegalStateException exception) {
-      LOG.info("Responding with status {}: {}", INTERNAL_SERVER_ERROR, exception.getMessage());
+      LOG.info("Responding with status {}: {}", INTERNAL_SERVER_ERROR, exception.getMessage(),
+          exception);
 
       ErrorResponse errorResponse = new ErrorResponse();
       errorResponse.setCode("InternalError");
