+++ server/src/test/java/com/adobe/testing/s3mock/ContextPathFileStoreControllerTest.java
@@ -28,6 +28,7 @@ import com.adobe.testing.s3mock.service.ObjectService;
 import com.adobe.testing.s3mock.store.BucketStore;
 import com.adobe.testing.s3mock.store.FileStore;
 import com.adobe.testing.s3mock.store.KmsKeyStore;
+import com.adobe.testing.s3mock.store.MultipartStore;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import java.nio.file.Paths;
@@ -46,8 +47,9 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 @AutoConfigureWebMvc
 @AutoConfigureMockMvc
-@MockBean(classes = {KmsKeyStore.class, FileStore.class, BucketStore.class, ObjectService.class})
-@SpringBootTest(classes = {S3MockConfiguration.class, MultipartService.class},
+@MockBean(classes = {KmsKeyStore.class, FileStore.class, BucketStore.class, ObjectService.class,
+    MultipartService.class, MultipartStore.class})
+@SpringBootTest(classes = {S3MockConfiguration.class},
     properties = {"com.adobe.testing.s3mock.contextPath=s3-mock"})
 class ContextPathFileStoreControllerTest {
   private static final Owner TEST_OWNER = new Owner(123, "s3-mock-file-store");
