+++ server/src/test/java/com/adobe/testing/s3mock/store/FileStoreTest.java
@@ -20,38 +20,24 @@ import static com.adobe.testing.s3mock.util.DigestUtil.hexDigest;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
-import static java.util.stream.Collectors.toList;
-import static java.util.stream.IntStream.rangeClosed;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.util.Files.contentOf;
 import static org.junit.jupiter.api.parallel.ExecutionMode.SAME_THREAD;
-import static org.springframework.http.MediaType.APPLICATION_OCTET_STREAM;
 
-import com.adobe.testing.s3mock.dto.CompletedPart;
-import com.adobe.testing.s3mock.dto.MultipartUpload;
-import com.adobe.testing.s3mock.dto.Owner;
-import com.adobe.testing.s3mock.dto.Part;
-import com.adobe.testing.s3mock.dto.Range;
 import com.adobe.testing.s3mock.dto.Tag;
 import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.ArrayUtils;
 import org.apache.http.entity.ContentType;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
@@ -63,44 +49,29 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 
 @AutoConfigureWebMvc
 @AutoConfigureMockMvc
-@MockBean(classes = {KmsKeyStore.class, BucketStore.class})
+@MockBean(classes = {KmsKeyStore.class, BucketStore.class, MultipartStore.class})
 @SpringBootTest(classes = {DomainConfiguration.class})
 @Execution(SAME_THREAD)
 class FileStoreTest {
-
   private static final String SIGNED_CONTENT =
       "24;chunk-signature=11707b33deb094881a16c70e9cbd5d79053a0bb235c25674e3cf0fed601683b5\r\n"
           + "## sample test file ##\n"
           + "\n"
           + "demo=content\n"
           + "0;chunk-signature=2206490f19c068b46367173d1e155b597fd367037fa3f924290b41c1e83c1c08";
-
   private static final String UNSIGNED_CONTENT =
       "## sample test file ##\n"
           + "\n"
           + "demo=content";
-
   private static final String TEST_BUCKET_NAME = "test-bucket";
-  private static final String ALL_BUCKETS = null;
-  private static final String NO_PREFIX = null;
-
   private static final String TEST_FILE_PATH = "src/test/resources/sampleFile.txt";
-
   private static final String NO_ENC = null;
   private static final String NO_ENC_KEY = null;
   private static final Map<String, String> NO_USER_METADATA = emptyMap();
   private static final String TEST_ENC_TYPE = "aws:kms";
-
   private static final String TEST_ENC_KEY = "aws:kms" + UUID.randomUUID();
-
-  private static final String DEFAULT_CONTENT_TYPE =
-      ContentType.APPLICATION_OCTET_STREAM.toString();
-
-  private static final Owner TEST_OWNER = new Owner(123, "s3-mock-file-store");
-
   private static final String TEXT_PLAIN = ContentType.TEXT_PLAIN.toString();
   private static final String ENCODING_GZIP = "gzip";
-
   private static final List<UUID> idCache = Collections.synchronizedList(new ArrayList<>());
 
   @Autowired
@@ -433,413 +404,6 @@ class FileStoreTest {
 
     assertThat(objectDeleted).as("Deletion should succeed!").isTrue();
     assertThat(s3ObjectMetadata).as("Object should be null!").isNull();
-  }
-
-  @Test
-  void shouldCreateMultipartUploadFolder() {
-    String fileName = "aFile";
-    String uploadId = "12345";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP,
-        uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    final File destinationFolder =
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, id.toString(), uploadId)
-            .toFile();
-
-    assertThat(destinationFolder.exists()).as("Destination folder does not exist").isTrue();
-    assertThat(destinationFolder.isDirectory()).as("Destination folder is not a directory")
-        .isTrue();
-
-    fileStore.abortMultipartUpload(metadataFrom(TEST_BUCKET_NAME), id, uploadId);
-  }
-
-  @Test
-  void shouldCreateMultipartUploadFolderIfBucketExists() {
-    String uploadId = "12345";
-    String fileName = "aFile";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    final File destinationFolder =
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, id.toString(), uploadId)
-            .toFile();
-
-    assertThat(destinationFolder.exists()).as("Destination folder does not exist").isTrue();
-    assertThat(destinationFolder.isDirectory()).as("Destination folder is not a directory")
-        .isTrue();
-
-    fileStore.abortMultipartUpload(metadataFrom(TEST_BUCKET_NAME), id, uploadId);
-  }
-
-  @Test
-  void shouldStorePart() {
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    final String partNumber = "1";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    fileStore.putPart(
-        metadataFrom(TEST_BUCKET_NAME), id, uploadId, partNumber,
-        new ByteArrayInputStream("Test".getBytes()), false, NO_ENC, NO_ENC_KEY);
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, id.toString(), uploadId,
-                partNumber + ".part")
-            .toFile()
-            .exists()).as("Part does not exist!").isTrue();
-
-    fileStore.abortMultipartUpload(metadataFrom(TEST_BUCKET_NAME), id, uploadId);
-  }
-
-  @Test
-  void shouldFinishUpload() {
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    fileStore
-        .putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "1",
-            new ByteArrayInputStream("Part1".getBytes()), false, NO_ENC, NO_ENC_KEY);
-    fileStore
-        .putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "2",
-            new ByteArrayInputStream("Part2".getBytes()), false, NO_ENC, NO_ENC_KEY);
-
-    final String etag =
-        fileStore.completeMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id, uploadId,
-            getParts(2),
-            NO_ENC, NO_ENC_KEY);
-    final byte[] allMd5s = ArrayUtils.addAll(
-        DigestUtils.md5("Part1"),
-        DigestUtils.md5("Part2")
-    );
-
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, id.toString(),
-                "fileData").toFile()
-            .exists()).as("File does not exist!").isTrue();
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, id.toString(),
-                "metadata").toFile()
-            .exists()).as("Metadata does not exist!").isTrue();
-    assertThat(etag).as("Special etag doesn't match.")
-        .isEqualTo(DigestUtils.md5Hex(allMd5s) + "-2");
-  }
-
-  @Test
-  void hasValidMetadata() {
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    fileStore
-        .putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "1",
-            new ByteArrayInputStream("Part1".getBytes()), false, NO_ENC, NO_ENC_KEY);
-    fileStore
-        .putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "2",
-            new ByteArrayInputStream("Part2".getBytes()), false, NO_ENC, NO_ENC_KEY);
-
-    fileStore.completeMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id, uploadId,
-        getParts(2),
-        NO_ENC, NO_ENC_KEY);
-
-    final S3ObjectMetadata s3ObjectMetadata =
-        fileStore.getS3Object(metadataFrom(TEST_BUCKET_NAME), id);
-    assertThat(s3ObjectMetadata.getSize()).as("Size doesn't match.").isEqualTo("10");
-    assertThat(s3ObjectMetadata.getContentType()).isEqualTo(APPLICATION_OCTET_STREAM.toString());
-  }
-
-  private List<CompletedPart> getParts(int n) {
-    List<CompletedPart> parts = new ArrayList<>();
-    for (int i = 1; i <= n; i++) {
-      CompletedPart part = new CompletedPart();
-      part.setPartNumber(i);
-      parts.add(part);
-    }
-    return parts;
-  }
-
-  @Test
-  void returnsValidPartsFromMultipart() {
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    UUID id = managedId();
-    String part1 = "Part1";
-    ByteArrayInputStream part1Stream = new ByteArrayInputStream(part1.getBytes());
-    String part2 = "Part2";
-    ByteArrayInputStream part2Stream = new ByteArrayInputStream(part2.getBytes());
-    final Part expectedPart1 = prepareExpectedPart(1, part1);
-    final Part expectedPart2 = prepareExpectedPart(2, part2);
-
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    fileStore.putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "1", part1Stream, false, NO_ENC,
-        NO_ENC_KEY);
-    fileStore.putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "2", part2Stream, false, NO_ENC,
-        NO_ENC_KEY);
-
-    List<Part> parts =
-        fileStore.getMultipartUploadParts(metadataFrom(TEST_BUCKET_NAME), id, uploadId);
-
-    assertThat(parts.size()).as("Part quantity does not match").isEqualTo(2);
-
-    expectedPart1.setLastModified(parts.get(0).getLastModified());
-    expectedPart2.setLastModified(parts.get(1).getLastModified());
-
-    assertThat(parts.get(0)).as("Part 1 attributes doesn't match").isEqualTo(expectedPart1);
-    assertThat(parts.get(1)).as("Part 2 attributes doesn't match").isEqualTo(expectedPart2);
-
-    fileStore.abortMultipartUpload(metadataFrom(TEST_BUCKET_NAME), id, uploadId);
-  }
-
-  private Part prepareExpectedPart(final int partNumber, final String content) {
-    Part part = new Part();
-    part.setETag(String.format("%s", DigestUtils.md5Hex(content)));
-    part.setPartNumber(partNumber);
-    part.setSize((long) content.getBytes().length);
-    return part;
-  }
-
-  @Test
-  void deletesTemporaryMultipartUploadFolder() {
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    fileStore
-        .putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "1",
-            new ByteArrayInputStream("Part1".getBytes()), false, NO_ENC, NO_ENC_KEY);
-
-    fileStore.completeMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id, uploadId,
-        getParts(1),
-        NO_ENC, NO_ENC_KEY);
-
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, fileName, uploadId)
-            .toFile()
-            .exists()).as("Folder should not exist anymore!").isFalse();
-  }
-
-  @Test
-  void listsMultipartUploads() {
-    assertThat(fileStore.listMultipartUploads(ALL_BUCKETS, NO_PREFIX)).isEmpty();
-
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    UUID id = managedId();
-    BucketMetadata bucketMetadata = metadataFrom(TEST_BUCKET_NAME);
-    final MultipartUpload initiatedUpload = fileStore
-        .prepareMultipartUpload(bucketMetadata, fileName, id, DEFAULT_CONTENT_TYPE, ENCODING_GZIP,
-            uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    final Collection<MultipartUpload> uploads = fileStore.listMultipartUploads(TEST_BUCKET_NAME,
-        NO_PREFIX);
-    assertThat(uploads).hasSize(1);
-    final MultipartUpload upload = uploads.iterator().next();
-    assertThat(upload).isEqualTo(initiatedUpload);
-    // and some specific sanity checks
-    assertThat(upload.getUploadId()).isEqualTo(uploadId);
-    assertThat(upload.getKey()).isEqualTo(fileName);
-
-    fileStore.completeMultipartUpload(bucketMetadata, fileName, id, uploadId, getParts(0),
-        NO_ENC, NO_ENC_KEY);
-
-    assertThat(fileStore.listMultipartUploads(ALL_BUCKETS, NO_PREFIX)).isEmpty();
-  }
-
-  @Test
-  void listsMultipartUploadsMultipleBuckets() {
-    assertThat(fileStore.listMultipartUploads(ALL_BUCKETS, NO_PREFIX)).isEmpty();
-
-    final String fileName1 = "PartFile1";
-    final String uploadId1 = "123451";
-    final String bucketName1 = "bucket1";
-    UUID id = managedId();
-    final MultipartUpload initiatedUpload1 = fileStore
-        .prepareMultipartUpload(metadataFrom(bucketName1), fileName1, id, DEFAULT_CONTENT_TYPE,
-            ENCODING_GZIP,
-            uploadId1, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    final String fileName2 = "PartFile2";
-    final String uploadId2 = "123452";
-    final String bucketName2 = "bucket2";
-    final MultipartUpload initiatedUpload2 = fileStore
-        .prepareMultipartUpload(metadataFrom(bucketName2), fileName2, id, DEFAULT_CONTENT_TYPE,
-            ENCODING_GZIP,
-            uploadId2, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    final Collection<MultipartUpload> uploads1 = fileStore.listMultipartUploads(bucketName1,
-        NO_PREFIX);
-    assertThat(uploads1).hasSize(1);
-    final MultipartUpload upload1 = uploads1.iterator().next();
-    assertThat(upload1).isEqualTo(initiatedUpload1);
-    // and some specific sanity checks
-    assertThat(upload1.getUploadId()).isEqualTo(uploadId1);
-    assertThat(upload1.getKey()).isEqualTo(fileName1);
-
-    final Collection<MultipartUpload> uploads2 = fileStore.listMultipartUploads(bucketName2,
-        NO_PREFIX);
-    assertThat(uploads2).hasSize(1);
-    final MultipartUpload upload2 = uploads2.iterator().next();
-    assertThat(upload2).isEqualTo(initiatedUpload2);
-    // and some specific sanity checks
-    assertThat(upload2.getUploadId()).isEqualTo(uploadId2);
-    assertThat(upload2.getKey()).isEqualTo(fileName2);
-
-    fileStore.completeMultipartUpload(metadataFrom(bucketName1), fileName1, id, uploadId1,
-        getParts(0),
-        NO_ENC, NO_ENC_KEY);
-    fileStore.completeMultipartUpload(metadataFrom(bucketName2), fileName2, id, uploadId2,
-        getParts(0),
-        NO_ENC, NO_ENC_KEY);
-
-    assertThat(fileStore.listMultipartUploads(ALL_BUCKETS, NO_PREFIX)).isEmpty();
-  }
-
-  @Test
-  void abortMultipartUpload() {
-    assertThat(fileStore.listMultipartUploads(ALL_BUCKETS, NO_PREFIX)).isEmpty();
-
-    final String fileName = "PartFile";
-    final String uploadId = "12345";
-    UUID id = managedId();
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), fileName, id,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    fileStore.putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, "1",
-        new ByteArrayInputStream("Part1".getBytes()), false, NO_ENC, NO_ENC_KEY);
-    assertThat(fileStore.listMultipartUploads(TEST_BUCKET_NAME, NO_PREFIX)).hasSize(1);
-
-    fileStore.abortMultipartUpload(metadataFrom(TEST_BUCKET_NAME), id, uploadId);
-
-    assertThat(fileStore.listMultipartUploads(ALL_BUCKETS, NO_PREFIX)).isEmpty();
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, fileName,
-                "fileData").toFile()
-            .exists()).as("File exists!").isFalse();
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, fileName,
-                "metadata").toFile()
-            .exists()).as("Metadata exists!").isFalse();
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, fileName, uploadId)
-            .toFile()
-            .exists()).as("Temp upload folder exists!").isFalse();
-  }
-
-  @Test
-  void copyPart() {
-    final String sourceFile = UUID.randomUUID().toString();
-    final String uploadId = UUID.randomUUID().toString();
-    UUID sourceId = managedId();
-    final String targetFile = UUID.randomUUID().toString();
-    final String partNumber = "1";
-    UUID destinationId = managedId();
-
-    final byte[] contentBytes = UUID.randomUUID().toString().getBytes();
-    fileStore.putS3Object(metadataFrom(TEST_BUCKET_NAME), sourceId, sourceFile,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP,
-        new ByteArrayInputStream(contentBytes), false,
-        NO_USER_METADATA, NO_ENC, NO_ENC_KEY, emptyList());
-
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), targetFile, destinationId,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    Range range = new Range(0, contentBytes.length);
-    fileStore.copyPart(
-        metadataFrom(TEST_BUCKET_NAME), sourceId, range, partNumber,
-        metadataFrom(TEST_BUCKET_NAME), destinationId, uploadId);
-    assertThat(
-        Paths.get(rootFolder.getAbsolutePath(), TEST_BUCKET_NAME, destinationId.toString(),
-                uploadId,
-                partNumber + ".part")
-            .toFile()
-            .exists()).as("Part does not exist!").isTrue();
-    fileStore.abortMultipartUpload(metadataFrom(TEST_BUCKET_NAME), destinationId, uploadId);
-  }
-
-  @Test
-  void copyPartNoRange() {
-    final String sourceFile = UUID.randomUUID().toString();
-    final String uploadId = UUID.randomUUID().toString();
-    UUID sourceId = managedId();
-    final String targetFile = UUID.randomUUID().toString();
-    final String partNumber = "1";
-    UUID destinationId = managedId();
-    final byte[] contentBytes = UUID.randomUUID().toString().getBytes();
-    BucketMetadata bucketMetadata = metadataFrom(TEST_BUCKET_NAME);
-    fileStore.putS3Object(bucketMetadata, sourceId, sourceFile, DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP,
-        new ByteArrayInputStream(contentBytes), false,
-        NO_USER_METADATA, NO_ENC, NO_ENC_KEY, emptyList());
-
-    fileStore.prepareMultipartUpload(bucketMetadata, targetFile, destinationId,
-        DEFAULT_CONTENT_TYPE,
-        ENCODING_GZIP, uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-
-    fileStore.copyPart(
-        bucketMetadata, sourceId, null, partNumber,
-        bucketMetadata, destinationId, uploadId);
-
-    assertThat(
-        Paths.get(bucketMetadata.getPath().toString(), destinationId.toString(),
-                uploadId,
-                partNumber + ".part")
-            .toFile()
-            .exists()).as("Part does not exist!").isTrue();
-    fileStore.abortMultipartUpload(bucketMetadata, destinationId, uploadId);
-  }
-
-  @Test
-  void missingUploadPreparation() {
-    Range range = new Range(0, 0);
-    IllegalStateException e = Assertions.assertThrows(IllegalStateException.class, () ->
-        fileStore.copyPart(
-            metadataFrom(TEST_BUCKET_NAME), UUID.randomUUID(), range, "1",
-            metadataFrom(TEST_BUCKET_NAME), UUID.randomUUID(), UUID.randomUUID().toString())
-    );
-
-    assertThat(e.getMessage()).isEqualTo("Missed preparing Multipart Request.");
-  }
-
-  @Test
-  void multipartUploadPartsAreSortedNumerically() throws IOException {
-    final String uploadId = UUID.randomUUID().toString();
-    final String filename = UUID.randomUUID().toString();
-    UUID id = managedId();
-
-    fileStore.prepareMultipartUpload(metadataFrom(TEST_BUCKET_NAME), filename, id, TEXT_PLAIN,
-        ENCODING_GZIP,
-        uploadId, TEST_OWNER, TEST_OWNER, NO_USER_METADATA);
-    for (int i = 1; i < 11; i++) {
-      final ByteArrayInputStream inputStream = new ByteArrayInputStream((i + "\n").getBytes());
-
-      fileStore.putPart(metadataFrom(TEST_BUCKET_NAME), id, uploadId, String.valueOf(i),
-          inputStream, false, NO_ENC, NO_ENC_KEY);
-    }
-    fileStore.completeMultipartUpload(metadataFrom(TEST_BUCKET_NAME), filename, id, uploadId,
-        getParts(10),
-        NO_ENC, NO_ENC_KEY);
-    final List<String> s = FileUtils
-        .readLines(fileStore.getS3Object(metadataFrom(TEST_BUCKET_NAME), id).getDataPath().toFile(),
-            "UTF8");
-
-    assertThat(s).contains(rangeClosed(1, 10).mapToObj(Integer::toString)
-        .collect(toList()).toArray(new String[] {}));
   }
 
   private BucketMetadata metadataFrom(String bucketName) {
