+++ server/src/test/java/com/adobe/testing/s3mock/store/KmsKeyStoreTest.java
@@ -22,9 +22,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.MockBeans;
 
-@MockBeans({@MockBean(classes = BucketStore.class), @MockBean(classes = FileStore.class)})
+@MockBean(classes = {BucketStore.class, FileStore.class, MultipartStore.class})
 @SpringBootTest(classes = DomainConfiguration.class)
 class KmsKeyStoreTest {
